refactor(nextjs): rename route /embed to /upload
