+++ templates/types/streaming/nextjs/app/api/chat/llamaindex/documents.ts
