+++ templates/types/streaming/nextjs/app/api/chat/upload/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { initSettings } from "../engine/settings";
-import { uploadDocument } from "./embeddings";
+import { uploadDocument } from "../llamaindex/documents";
 
 initSettings();
